requirements.txt
django==3.2.4
djangorestframework==3.12.4
djangorestframework_simplejwt==4.7.1
django-filter==2.4.0
celery[redis]
factory_boy==3.2.0
drf-yasg==1.20.0
psycopg2-binary==2.9.1
coverage==5.5
flake8==3.9.2
flake8-docstrings==1.6.0
flake8-quotes==3.2.0
isort==5.9.1
pylint==2.8.3
pylint-django==2.0.8
